+++ macros/src/doc_string.rs
@@ -1,0 +1,53 @@
+use proc_macro2::TokenStream;
+use quote::{quote, ToTokens};
+use syn::parse::{Parse, ParseStream};
+use syn::{Attribute, Error, Expr, ExprLit, Lit, LitStr, Meta, MetaNameValue, Result};
+
+pub(crate) struct DocString {
+    inner: Vec<LitStr>,
+}
+
+impl DocString {
+    pub(crate) fn new() -> Self {
+        Self { inner: Vec::new() }
+    }
+
+    pub(crate) fn push(&mut self, attr: Attribute) -> Result<()> {
+        if !attr.path().is_ident("doc") {
+            return Err(Error::new_spanned(attr, "extected attribute `doc`"));
+        }
+
+        if let Meta::NameValue(MetaNameValue {
+            value: Expr::Lit(ExprLit {
+                lit: Lit::Str(doc), ..
+            }),
+            ..
+        }) = attr.meta
+        {
+            self.inner.push(doc);
+            Ok(())
+        } else {
+            Err(Error::new_spanned(attr, "only doc comments are supported"))
+        }
+    }
+}
+
+impl Parse for DocString {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let mut doc = Self::new();
+
+        let attrs = input.call(Attribute::parse_outer)?;
+
+        for attr in attrs {
+            doc.push(attr)?;
+        }
+
+        Ok(doc)
+    }
+}
+
+impl ToTokens for DocString {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        tokens.extend(self.inner.iter().map(|doc| quote!(#[doc = #doc])));
+    }
+}