+++ macros/src/service/ev_item.rs
@@ -1,4 +1,5 @@
 use super::{kw, Options};
+use crate::doc_string::DocString;
 use heck::ToUpperCamelCase;
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
@@ -8,6 +9,7 @@ use syn::parse::{Parse, ParseStream};
 use syn::{Ident, LitInt, Result, Token, Type};
 
 pub(super) struct EvItem {
+    doc: DocString,
     ident: Ident,
     ident_val: Ident,
     ident_ref: Ident,
@@ -36,8 +38,10 @@ impl EvItem {
         let krate = options.krate();
         let subscribe = &self.subscribe;
         let id = &self.id;
+        let doc = &self.doc;
 
         quote! {
+            #doc
             pub async fn #subscribe(&self) -> ::std::result::Result<(), #krate::Error> {
                 self.inner.subscribe(#id).await
             }
@@ -48,8 +52,10 @@ impl EvItem {
         let krate = options.krate();
         let unsubscribe = &self.unsubscribe;
         let id = &self.id;
+        let doc = &self.doc;
 
         quote! {
+            #doc
             pub async fn #unsubscribe(&self) -> ::std::result::Result<(), #krate::Error> {
                 self.inner.unsubscribe(#id).await
             }
@@ -84,13 +90,17 @@ impl EvItem {
     pub(crate) fn gen_variant(&self, options: &Options) -> TokenStream {
         let krate = options.krate();
         let variant = &self.variant;
+        let doc = &self.doc;
 
         let ty = match self.ty {
             Some(ref ty) => quote! { #ty },
             None => quote! { () },
         };
 
-        quote! { #variant(#krate::Event<#ty>), }
+        quote! {
+            #doc
+            #variant(#krate::Event<#ty>),
+        }
     }
 
     pub(crate) fn gen_emitters(&self, options: &Options) -> TokenStream {
@@ -99,9 +109,11 @@ impl EvItem {
         let ident_val = &self.ident_val;
         let ident_ref = &self.ident_ref;
         let id = &self.id;
+        let doc = &self.doc;
 
         if let Some(ref ty) = self.ty {
             quote! {
+                #doc
                 pub fn #ident<T>(&self, args: T) -> ::std::result::Result<(), #krate::Error>
                 where
                     T: #krate::core::Serialize<#krate::core::tags::As<#ty>>,
@@ -109,16 +121,19 @@ impl EvItem {
                     self.inner.emit_as::<#krate::core::tags::As<#ty>, _>(#id, args)
                 }
 
+                #doc
                 pub fn #ident_val(&self, args: #ty) -> ::std::result::Result<(), #krate::Error> {
                     self.#ident(args)
                 }
 
+                #doc
                 pub fn #ident_ref(&self, args: &#ty) -> ::std::result::Result<(), #krate::Error> {
                     self.#ident(args)
                 }
             }
         } else {
             quote! {
+                #doc
                 pub fn #ident(&self) -> ::std::result::Result<(), #krate::Error> {
                     self.inner.emit(#id, ())
                 }
@@ -155,6 +170,7 @@ impl EvItem {
 
 impl Parse for EvItem {
     fn parse(input: ParseStream) -> Result<Self> {
+        let doc = input.parse()?;
         input.parse::<kw::event>()?;
         let ident = input.parse::<Ident>()?;
         input.parse::<Token![@]>()?;
@@ -180,6 +196,7 @@ impl Parse for EvItem {
         );
 
         Ok(Self {
+            doc,
             ident,
             ident_val,
             ident_ref,