+++ macros/src/service/fn_fallback_item.rs
@@ -1,3 +1,4 @@
+use crate::doc_string::DocString;
 use heck::ToUpperCamelCase;
 use proc_macro2::TokenStream;
 use quote::quote;
@@ -6,6 +7,7 @@ use syn::parse::{Parse, ParseStream};
 use syn::{Ident, Result, Token, Type};
 
 pub(super) struct FnFallbackItem {
+    doc: DocString,
     ident: Ident,
     variant: Ident,
     ty: Type,
@@ -23,8 +25,10 @@ impl FnFallbackItem {
     pub(crate) fn gen_variant(&self) -> TokenStream {
         let variant = &self.variant;
         let ty = &self.ty;
+        let doc = &self.doc;
 
         quote! {
+            #doc
             #[allow(dead_code)]
             #variant(#ty),
         }
@@ -51,6 +55,7 @@ impl FnFallbackItem {
 
 impl Parse for FnFallbackItem {
     fn parse(input: ParseStream) -> Result<Self> {
+        let doc = input.parse()?;
         input.parse::<Token![fn]>()?;
         let ident = input.parse::<Ident>()?;
         input.parse::<Token![=]>()?;
@@ -62,6 +67,11 @@ impl Parse for FnFallbackItem {
             ident.span(),
         );
 
-        Ok(Self { ident, variant, ty })
+        Ok(Self {
+            doc,
+            ident,
+            variant,
+            ty,
+        })
     }
 }