+++ macros/src/service/fn_item.rs
@@ -1,4 +1,5 @@
 use super::{FnBody, Options};
+use crate::doc_string::DocString;
 use heck::ToUpperCamelCase;
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
@@ -9,6 +10,7 @@ use syn::token::Brace;
 use syn::{braced, Ident, LitInt, Result, Token, Type};
 
 pub(super) struct FnItem {
+    doc: DocString,
     ident: Ident,
     ident_val: Ident,
     ident_ref: Ident,
@@ -32,6 +34,7 @@ impl FnItem {
         let ident_val = &self.ident_val;
         let ident_ref = &self.ident_ref;
         let id = &self.id;
+        let doc = &self.doc;
 
         let ok = match self.body.ok() {
             Some(ok) => quote! { #ok },
@@ -45,6 +48,7 @@ impl FnItem {
 
         if let Some(args) = self.body.args() {
             quote! {
+                #doc
                 pub fn #ident<T>(&self, args: T) -> #krate::PendingReply<#ok, #err>
                 where
                     T: #krate::core::Serialize<#krate::core::tags::As<#args>>,
@@ -57,16 +61,19 @@ impl FnItem {
                         ).cast()
                 }
 
+                #doc
                 pub fn #ident_val(&self, args: #args) -> #krate::PendingReply<#ok, #err> {
                     self.#ident(args)
                 }
 
+                #doc
                 pub fn #ident_ref(&self, args: &#args) -> #krate::PendingReply<#ok, #err> {
                     self.#ident(args)
                 }
             }
         } else {
             quote! {
+                #doc
                 pub fn #ident(&self) -> #krate::PendingReply<#ok, #err> {
                     self.inner.call(#id, (), ::std::option::Option::Some(Self::VERSION)).cast()
                 }
@@ -77,6 +84,7 @@ impl FnItem {
     pub(crate) fn gen_variant(&self, options: &Options) -> TokenStream {
         let krate = options.krate();
         let variant = &self.variant;
+        let doc = &self.doc;
 
         let args = match self.body.args() {
             Some(args) => quote! { #args },
@@ -94,6 +102,7 @@ impl FnItem {
         };
 
         quote! {
+            #doc
             #variant(#krate::Call<#args, #ok, #err>),
         }
     }
@@ -178,6 +187,7 @@ impl FnItem {
 
 impl Parse for FnItem {
     fn parse(input: ParseStream) -> Result<Self> {
+        let doc = input.parse()?;
         input.parse::<Token![fn]>()?;
         let ident = input.parse::<Ident>()?;
         input.parse::<Token![@]>()?;
@@ -208,6 +218,7 @@ impl Parse for FnItem {
         );
 
         Ok(Self {
+            doc,
             ident,
             ident_val,
             ident_ref,